requirements.txt
tensorflow
git+https://github.com/ipazc/mtcnn.git
deepface
pymongo
opencv-python
imgaug
flask
python-dotenv
gdown
dnspython